[debug-info] Emit DW_AT_byte_size for non-default pointer/reference types
In the CHERI and Arm Morello targets we support a compilation mode that
has multiple pointer sizes (32/64-bit integers and capabilities that are
twice that size). In order to support this in the debugger we emit
DW_AT_byte_size for pointer and reference types if their size is different
from the target's address space 0 pointer size.

Differential Revision: https://reviews.llvm.org/D99411