[llvm-ranlib] Handle -D and -U command line flag
Summary:
I have been trying to build CheriBSD (a fork for FreeBSD for the CHERI
CPU) with LLVM binutils instead of the default elftoolchain utilities.
I noticed that building static archives was failing because ranlib is
invoked with the -D flag. This failed with llvm-ranlib since it parses
the -D flag as the archive path and reports an error that more than one
archive has been passed.
This change also adjusts the short option parsing to handle the -h/-v flag
as part of the combined short options

This fixes https://llvm.org/PR41707

Reviewers: rupprecht

Subscribers: emaste, krytarowski, cfe-commits, llvm-commits

Tags: #clang, #llvm

Differential Revision: https://reviews.llvm.org/D71554