[libcxx][executors] Add a --test-executables flag
By explicitly saying which of the arguments is the test executable, we
can use that to codesign the test executables rather than commandLine[0].

This makes it possible to run tests under GDB/LLDB/Valgrind/qemu-user etc.
by prepending a wrapper command. A follow-up change will use this flag to
enable running all tests under GDB (printing a backtrace on crashes).

Differential Revision: https://reviews.llvm.org/D88899