[lit][libcxx] Add a --run-with-debugger/--gdb flag
This flag can be used to run libc++/libunwind/libc++abi tests under GDB to
make it easier to debug test failures. We use this approach in our CHERI
Jenkins setup to ensure that crashing test cases include a backtrace in
the Jenkins JUnit test result viewer.
This uses the --wrapper-command executor flag added in D88899.

Differential Revision: https://reviews.llvm.org/D88903